Factored out the MenuBar from the BlackjackGui, and wrote a controller for it as well to contain actionlisteners that were originally in blackjackguicontroller
